Fix a issue in undefineSecret.py
When usage type is volume, image need to be removed after undefine
secret.

Reviewed-on: https://code.engineering.redhat.com/gerrit/145858
Tested-by: Libvirt Jenkins <libvirt-jenkins@redhat.com>

+++ repos/secret/undefineSecret.py
@@ -6,7 +6,7 @@ from src import sharedmod
 from xml.dom import minidom
 
 required_params = ('secretUUID',)
-optional_params = {}
+optional_params = {'usagetype': 'volume'}
 
 
 def check_undefineSecret(ephemeral, secretUUID):
@@ -31,19 +31,23 @@ def undefineSecret(params):
     global conn
     logger = params['logger']
     secretUUID = params['secretUUID']
+    usagetype = params.get('usagetype', 'volume')
+
     conn = sharedmod.libvirtobj['conn']
 
     secretobj = conn.secretLookupByUUIDString(secretUUID)
     ephemeral = minidom.parseString(secretobj.XMLDesc(0)).\
         getElementsByTagName('secret')[0].getAttribute('ephemeral')
-    diskpath = minidom.parseString(secretobj.XMLDesc(0)).\
-        getElementsByTagName('volume')[0].childNodes[0].data
+    if usagetype == "volume":
+        diskpath = minidom.parseString(secretobj.XMLDesc(0)).\
+            getElementsByTagName('volume')[0].childNodes[0].data
     try:
         secretobj.undefine()
         if check_undefineSecret(ephemeral, secretUUID):
             logger.info("undefine the secret %s is successful" % secretUUID)
-            logger.info("remove the related volume %s" % diskpath)
-            os.remove(diskpath)
+            if usagetype == "volume":
+                logger.info("remove the related volume %s" % diskpath)
+                os.remove(diskpath)
         else:
             logger.error("fail to check secret undefine")
             return 1
